data: Split file hunk cache writes
File hunk rendering and selected-change persistence lived together in
data.file_hunk_display. The rendering module now only builds file-scoped
line views.

Cache writes update selected kind, patch lines, processed IDs, line-change
JSON, and snapshots. That persistence role belongs beside other
selected-change state.

This commit extracts cache_file_as_single_hunk and
cache_unstaged_file_as_single_hunk to data.selected_change.file_hunk_cache
and updates callers by cache versus render use.

File-scoped hunk rendering and selected-change cache persistence now have
separate modules.
