+++ src/git_stage_batch/commands/file_scope/include_file_replacement.py
@@ -11,7 +11,7 @@ from ...data.file_change_status import (
     file_has_staged_changes,
     file_has_unstaged_changes,
 )
-from ...data.file_hunk_display import cache_unstaged_file_as_single_hunk
+from ...data.selected_change.file_hunk_cache import cache_unstaged_file_as_single_hunk
 from ...data.line_id_files import write_line_ids_file
 from ...data.file_review.state import clear_last_file_review_state_if_file_matches
 from ...data.file_tracking import auto_add_untracked_files
