+++ src/git_stage_batch/data/file_hunk_display.py
@@ -1,10 +1,9 @@
-"""File-scoped text hunk rendering and selected-state caching."""
+"""File-scoped text hunk rendering."""
 
 from __future__ import annotations
 
 from collections.abc import Generator
 import os
-import json
 import subprocess
 import tempfile
 from typing import Optional
@@ -13,7 +12,6 @@ from ..core.diff_parser import (
     acquire_unified_diff,
     build_line_changes_from_patch_lines,
 )
-from ..core.hashing import compute_stable_hunk_hash_from_lines
 from ..core.models import (
     BinaryFileChange,
     GitlinkChange,
@@ -24,102 +22,13 @@ from ..core.models import (
     TextFileDeletionChange,
 )
 from ..core.buffer import LineBuffer
-from .line_id_files import write_line_ids_file
 from ..utils.repository_buffers import load_git_object_as_buffer
 from ..i18n import ngettext
-from ..utils.file_io import write_text_file_contents
 from ..utils.git_command import stream_git_command
-from ..utils.paths import (
-    get_context_lines,
-    get_line_changes_json_file_path,
-    get_processed_include_ids_file_path,
-    get_processed_skip_ids_file_path,
-    get_selected_hunk_hash_file_path,
-)
+from ..utils.paths import get_context_lines
 from ..core.text_lines import bytes_to_lines
 from .file_tracking import auto_add_untracked_files
-from .line_state import convert_line_changes_to_serializable_dict
 from .live_diff import stream_live_git_diff
-from .selected_change.store import (
-    SelectedChangeKind,
-    write_selected_change_kind,
-    write_selected_hunk_patch_lines,
-)
-from .selected_change.snapshots import write_snapshots_for_selected_file_path
-
-
-def cache_file_as_single_hunk(file_path: str) -> Optional[LineLevelChange]:
-    """Cache all changes for a file as a single concatenated hunk.
-
-    Reads the CURRENT working tree state for the file and fetches ALL
-    hunks (ignoring blocklist/batches), concatenating them into one
-    LineLevelChange with continuous line IDs.
-
-    This always reflects the live working tree state, unlike regular
-    hunk caching which uses snapshots.
-
-    Args:
-        file_path: Repository-relative path to file
-
-    Returns:
-        LineLevelChange with all file changes, or None if no changes
-    """
-    try:
-        combined_line_changes = render_file_as_single_hunk(file_path)
-        return _cache_combined_file_line_changes(file_path, combined_line_changes)
-    except subprocess.CalledProcessError:
-        return None
-
-
-def cache_unstaged_file_as_single_hunk(file_path: str) -> Optional[LineLevelChange]:
-    """Cache the remaining unstaged changes for a file as a single hunk."""
-    try:
-        combined_line_changes = render_unstaged_file_as_single_hunk(file_path)
-        return _cache_combined_file_line_changes(file_path, combined_line_changes)
-    except subprocess.CalledProcessError:
-        return None
-
-
-def _cache_combined_file_line_changes(
-    file_path: str,
-    combined_line_changes: Optional[LineLevelChange],
-) -> Optional[LineLevelChange]:
-    """Persist a combined file-scoped view as the current selection."""
-    if combined_line_changes is None:
-        return None
-
-    patch_lines = [
-        f"--- a/{file_path}\n".encode("utf-8"),
-        f"+++ b/{file_path}\n".encode("utf-8"),
-        (
-            f"@@ -{combined_line_changes.header.old_start},{combined_line_changes.header.old_len} "
-            f"+{combined_line_changes.header.new_start},{combined_line_changes.header.new_len} @@\n"
-        ).encode("utf-8"),
-    ]
-    for entry in combined_line_changes.lines:
-        patch_lines.append(entry.kind.encode("utf-8") + entry.text_bytes + b"\n")
-        if not entry.has_trailing_newline:
-            patch_lines.append(b"\\ No newline at end of file\n")
-
-    patch_hash = compute_stable_hunk_hash_from_lines(patch_lines)
-
-    write_selected_hunk_patch_lines(patch_lines)
-    write_text_file_contents(get_selected_hunk_hash_file_path(), patch_hash)
-    write_selected_change_kind(SelectedChangeKind.FILE)
-    write_line_ids_file(get_processed_include_ids_file_path(), set())
-    write_line_ids_file(get_processed_skip_ids_file_path(), set())
-    write_text_file_contents(
-        get_line_changes_json_file_path(),
-        json.dumps(
-            convert_line_changes_to_serializable_dict(combined_line_changes),
-            ensure_ascii=False,
-            indent=0,
-        ),
-    )
-
-    write_snapshots_for_selected_file_path(file_path)
-
-    return combined_line_changes
 
 
 def render_file_as_single_hunk(file_path: str) -> Optional[LineLevelChange]:
